Timer function timer_proc needs to serialize MiniportTimer calls, but all timers in ndiswrapper are implemented with one timer interface (wrap_timer). Instead of checking if a driver is serialized in timer_proc, check and set wrap_timer->wd during initialization (wrap_init_timer). Since we don't use wrap_timer->wd for any other purpose, this should be fine.
Note: In Linux, spinlocks are initialized to 1 and every time lock is tried,
it is decremented by one. This is easier to implement. However, I noticed
that with that implementation, some drivers seemed to call spinlock function
with uninitialized spinlocks - they were set to 0, instead of 1. I now
realize that this may be because we didn't initialize wnd->nmb->lock until
recently. So it is possible that those drivers (I don't remember which, but
have a vague recollection it may be acx100) may have been serialized drivers,
passing wnd->nmb->lock. It would be a good idea to investigate this and
try if using Linux implementation of spinlocks would work now.