Use del_timer_sync to cancel timer instead of del_timer. It is okay to use del_timer_sync, as (Ke/Ndis)CancelTimer can be called at <= DISPATCH_LEVEL and del_timer_sync shouldn't be called at interrupt context.
Use wrap_timer as data for timer_proc, so that if nt_timer is corrupted
for some reason, timers won't be affected (kdpc associated with nt_timer
may be corrupted and may result in a crash, though).

To keep track of number of entries in SList, use 'depth' field in union
to simplify SList implementation.