When loading driver, allocate space for bin files only if there are non-zero number of such files (most drivers don't have them); otherwise, calling kmalloc with 0 will cause slub allocator to grumble (this was noted couple of commits ago, with wrong comment about what was causing it).
Windows uses millisecond / 100 nanosecond granularity for timers. When they
are converted to HZ, we add 1, as the converted value may be 1 less than
the expiration timer, which I thought could be a problem, especially if HZ
is too low (e.g., 101 ms is converted to 25 HZ). However, this conservative
conversion _seems_ to cause problems with at least mrv8000c driver: Right
after MiniportInitialize completes, MiniportTimer is called a bit later than
what the driver expects, which causes crash sometimes. The fix _seems_ to
be not to add 1 to the converted value. More testing (with other drivers)
is necessary to make sure this doesn't break other drivers.

Use atomic insert/remove list macros to chain wrap_timers instead of using
spinlock.

Annotated some of the functions with assertions about IRQL at which they
should be running. More annotations should be added, so any issues become
clear as early as possible with as much information as possible.