Till now, we have been emulating DISPATCH_LEVEL IRQL by disabling preemption (and earlier, even bottom halves). This is problematic with either preempt disabled or RT-preempt is used. With this patch, preemption is emulated by mutex so that the thread holding the mutex is running at DISPATCH_LEVEL and other threads that want to enter DISPATCH_LEVEL must wait. But Linux doesn't have recursive mutex, so if a thread that is already running at DISPATCH_LEVEL (hodling the mutex) wants to enter DISPATCH_LEVEL, we shouldn't wait for the mutex (otherwise, it would cause deadlock).
Till now, NT spinlocks are used almost everywhere. Instead, use Linux's
spinlocks where possible. This makes it easier to debug ndiswrapper (through
Linux's debug facilities).

This is rather intrusive patch, so likely to break things (I may even have
to revert it if necessary if it stands in the way of next release).