Fix occasional scan failures - try reading scan data 10 times
The required space may grow between requests, so be prepared that two
tries may be insufficient.

Sanitize the logic in iw_get_scan().  Only reallocate the buffer and
retry if more space is requested, regardless of the error code.

Check the amount of data actually written by the driver.  Drop items
that don't fit the written data.