Make win2lin stubs appear in backtrace under full function names
Instead of going to the common code with jmp, use macros to expand the
code for every function.  This increases the code by about 8k (when
stripped), but it greatly increases usefulness of the backtraces.

Considering that the code is for x86_64 only and that ndiswrapper is not
exactly stable yet, it appears to be a good trade-off.