Improve sanity checks
Check for .config before checking for the version header.  If we have
the former, but not the later, ask user to run "make modules_prepare".

Not finding UTS_RELEASE in an existing version header has nothing to do
with "make modules_prepare".  It's more likely that our build system
doesn't support some future kernel.

Remove the second check for the version header, it should never be
triggered.